pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "cv-scribble-diffusion"
version = "0.1.0"
description = "Real-time OpenCV scribble-to-image generation with ControlNet and Stable Diffusion."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"numpy>=1.24,<3",
	"opencv-python>=4.8,<5",
	"pillow>=10,<12",
	"diffusers>=0.25,<0.35",
	"torch>=2.1,<3",
	"accelerate>=0.25,<2",
]

[project.optional-dependencies]
dev = [
	"pytest>=7,<9",
	"pytest-cov>=4,<6",
]

[project.scripts]
cv-scribble-diffusion = "cv_scribble_diffusion.app.startup:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
